parse-balance-map extracted in it's own file for erc1155
