+++ packages/drop-sdk/src/index.ts
@@ -1,4 +1,4 @@
-import parseBalanceMap, { MerkleDistributorInfoERC1155, RecipientsDictFormatERC1155, RecipientsArrayFormatERC1155 } from "./merkle/parse-balance-map"
+import parseBalanceMap, { MerkleDistributorInfoERC1155, RecipientsDictFormatERC1155, RecipientsArrayFormatERC1155 } from "./merkle/parse-balance-map-erc1155"
 
 export const buildMerkleTreeERC1155 = (recipientsData: RecipientsDictFormatERC1155 | RecipientsArrayFormatERC1155[]): MerkleDistributorInfoERC1155 => {
     const merkleData = parseBalanceMap(recipientsData);
@@ -10,4 +10,4 @@ export const buildMerkleTreeERC1155 = (recipientsData: RecipientsDictFormatERC11
 //     return merkleData
 // }
 
-export { MerkleDistributorInfoERC1155, RecipientsDictFormatERC1155, RecipientsArrayFormatERC1155 } from "./merkle/parse-balance-map"
+export { MerkleDistributorInfoERC1155, RecipientsDictFormatERC1155, RecipientsArrayFormatERC1155 } from "./merkle/parse-balance-map-erc1155"
